+++ prism-agent/service/server/src/main/scala/io/iohk/atala/agent/server/Main.scala
@@ -3,6 +3,8 @@ package io.iohk.atala.agent.server
 import zio.*
 import io.iohk.atala.mercury.PeerDID
 import io.iohk.atala.mercury.AgentCli
+import io.iohk.atala.castor.sql.repository.{Migrations => CastorMigrations}
+import io.iohk.atala.pollux.sql.repository.{Migrations => PolluxMigrations}
 
 object Main extends ZIOAppDefault {
   override def run: ZIO[Any, Throwable, Unit] =
@@ -35,35 +37,42 @@ object Main extends ZIOAppDefault {
         case Some(s) if s.toIntOption.isDefined => s.toInt
         case _                                  => 8090
       }
-      _ <- ZIO.logInfo(s"DIDComm Service port => $didCommServicePort") 
-      
-//      COMMENTED IN ORDER TO RUN THE AGENT LOCALLY
-//      agentDID <- for {
-//        peer <- ZIO.succeed(PeerDID.makePeerDid(serviceEndpoint = Some(s"http://localhost:$didCommServicePort")))
-//        _ <- ZIO.logInfo(s"New DID: ${peer.did}") *>
-//          ZIO.logInfo(s"JWK for KeyAgreement: ${peer.jwkForKeyAgreement.toJSONString}") *>
-//          ZIO.logInfo(s"JWK for KeyAuthentication: ${peer.jwkForKeyAuthentication.toJSONString}")
-//      } yield (peer)
-//
-//      didCommLayer = AgentCli.agentLayer(agentDID)
-//
-//      didCommExchangesFiber <- Modules.didCommExchangesJob
-//        .provide(
-//          didCommLayer
-//        )
-//        .debug
-//        .fork
-//
-//      didCommServiceFiber <- Modules
-//        .didCommServiceEndpoint(didCommServicePort)
-//        .provide(
-//          didCommLayer,
-//          AppModule.credentialServiceLayer
-//        )
-//        .debug
-//        .fork
+      _ <- ZIO.logInfo(s"DIDComm Service port => $didCommServicePort")
 
-      _ <- Modules.app(restServicePort).fork
+      // Execute migrations from Castor and Pollux libraries using Flyway
+      _ <- ZIO
+        .serviceWithZIO[CastorMigrations](_.migrate)
+        .provide(RepoModule.castorDbConfigLayer >>> CastorMigrations.layer)
+      _ <- ZIO
+        .serviceWithZIO[PolluxMigrations](_.migrate)
+        .provide(RepoModule.polluxDbConfigLayer >>> PolluxMigrations.layer)
+
+      agentDID <- for {
+        peer <- ZIO.succeed(PeerDID.makePeerDid(serviceEndpoint = Some(s"http://localhost:$didCommServicePort")))
+        _ <- ZIO.logInfo(s"New DID: ${peer.did}") *>
+          ZIO.logInfo(s"JWK for KeyAgreement: ${peer.jwkForKeyAgreement.toJSONString}") *>
+          ZIO.logInfo(s"JWK for KeyAuthentication: ${peer.jwkForKeyAuthentication.toJSONString}")
+      } yield (peer)
+
+      didCommLayer = AgentCli.agentLayer(agentDID)
+
+      didCommExchangesFiber <- Modules.didCommExchangesJob
+        .provide(
+          didCommLayer
+        )
+        .debug
+        .fork
+
+      didCommServiceFiber <- Modules
+        .didCommServiceEndpoint(didCommServicePort)
+        .provide(
+          didCommLayer,
+          AppModule.credentialServiceLayer
+        )
+        .debug
+        .fork
+
+      _ <- Modules.app(restServicePort)
       _ <- Modules.zioApp
       _ <- ZIO.never
     } yield ()